requirements.txt
requests>=2.31.0
pandas>=2.2.0
polars>=0.20.0
pyarrow>=16.0.0
geopandas>=0.14.0
shapely>=2.0.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=5.0.0
requests-mock>=1.9.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
folium>=0.14.0
plotly>=5.17.0
jupyter>=1.0.0
notebook>=7.0.0
ipython>=8.0.0
